Move persistent cache creation to separate file so we can run gc on build
