+++ lib/graph-script.js
@@ -3,10 +3,7 @@ var concat = require('concat-stream')
 var to = require('flush-write-stream')
 var tfilter = require('tfilter')
 var assert = require('assert')
-var path = require('path')
-var os = require('os')
 
-var browserifyPersistFs = require('browserify-persist-fs')
 var babelPresetEnv = require('babel-preset-env')
 var splitRequire = require('split-require')
 var cssExtract = require('css-extract')
@@ -19,6 +16,7 @@ var tinyify = require('tinyify')
 var glslify = require('glslify')
 var brfs = require('brfs')
 
+var createPersistentCache = require('./persistent-cache')
 var ttyError = require('./tty-error')
 var exorcise = require('./exorcise')
 
@@ -47,9 +45,7 @@ function node (state, createEdge) {
   var self = this
   var entry = state.metadata.entry
   var fullPaths = Boolean(state.metadata.fullPaths)
-  var cacheDir = path.join(os.homedir(), '.cache', 'bankai')
-  var persistentCache = browserifyPersistFs(cacheDir, getCacheId(state))
-  var b = browserify(browserifyOpts([entry], fullPaths, persistentCache))
+  var b = browserify(browserifyOpts([entry], fullPaths, createPersistentCache(state)))
   var shouldMinify = !state.metadata.watch
 
   if (state.metadata.watch) {
@@ -171,29 +167,5 @@ function browserifyOpts (entries, fullPaths, persistentCache) {
     packageCache: {},
     cache: {},
     persistentCache: persistentCache
-  }
-}
-
-function getCacheId (state) {
-  return {
-    watch: state.metadata.watch,
-    entry: state.metadata.entry,
-    transforms: [
-      'sheetify',
-      'glslify',
-      'tfilter',
-      'brfs',
-      'yo-yoify',
-      'css-extract',
-      'tinyify',
-      'split-require',
-      // TODO should probably add custom .babelrc here too if we can
-      'babelify',
-      'babel-preset-env'
-    ].map(version)
-  }
-
-  function version (mod) {
-    return require(mod + '/package.json').version
   }
 }
